src/test/java/interviews/microsoft/noodleEssence/_340_Longest_Substring_with_At_Most_K_Distinct_Characters/SolutionTest.java
package interviews.microsoft.noodleEssence._340_Longest_Substring_with_At_Most_K_Distinct_Characters;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class SolutionTest {
  private Solution solution;

  @BeforeEach
  void setUp() throws Exception {
    solution = new Solution();
  }

  @Test
  void testOnlineCase() throws Exception {
    final String s = "eceba";
    assertEquals(3, solution.lengthOfLongestSubstringKDistinct(s, 2));
  }

  @Test
  void testKEquals0() throws Exception {
    final String s = "abaccc";
    assertEquals(4, solution.lengthOfLongestSubstringKDistinct(s, 2));
  }
}